Refactor the routes and names related with the rentals
