+++ src/server/server.js
@@ -11,6 +11,7 @@ const appRouter = require('./routers/app');
 const locationsRouter = require('./routers/locations');
 const lockersRouter = require('./routers/lockers');
 const usersRouter = require('./routers/users');
+const rentalsRouter = require('./routers/rentals');
 const config = require('./config/server.json')[env];
 const { globalErrorHandler, NotFoundError } = require('./errors');
 
@@ -52,6 +53,7 @@ app.use('/api/v1/admin', adminRouter);
 app.use('/api/v1/locations', locationsRouter);
 app.use('/api/v1/lockers', lockersRouter);
 app.use('/api/v1/users', usersRouter);
+app.use('/api/v1/rentals', rentalsRouter);
 app.use('/api/v1', router);
 app.use('/api/v1/*', () => {
 	throw new NotFoundError();
